move is-react check into utils to sort out that this is not meant to use in production
