+++ packages/dnb-ui-lib/src/shared/utils/is-react.js
